+++ src/MapperQuery/AbstractQueryParam.php
@@ -1,9 +1,11 @@
 <?php
 namespace Pyncer\Data\MapperQuery;
 
+use Pyncer\Exception\InvalidArgumentException;
+
 use function trim;
 
-abstract class AbstractQueryParam
+abstract class AbstractQueryParam implements QueryParamInterface
 {
     protected string $queryParamString;
     private ?array $parts = null;
@@ -27,8 +29,76 @@ abstract class AbstractQueryParam
         return $this;
     }
 
-    abstract public function addQueryParamString(string $value): static;
-    abstract public function getCleanQueryParamString(): string;
+    public function getCleanQueryParamString(): string
+    {
+        return $this->buildQueryParamStringFromParts($this->getParts());
+    }
+
+    public function addQueryParam(QueryParamInterface $value): static
+    {
+        $class = static::class;
+        if (!($value instanceof $class)) {
+           throw new InvalidArgumentException('Query param must be an instance of the same class.');
+        }
+
+        $this->queryParamString = $this->mergeQueryParamStrings(
+            $this->queryParamString,
+            $value->getQueryParamString()
+        );
+
+        $this->parts = null;
+
+        if ($this->hasCleaned() || $value->hasCleaned()) {
+            $this->cleanParts = $this->mergeQueryParamParts(
+                $this->getParts(),
+                $value->getParts(),
+            );
+        } else {
+            $this->cleanParts = null;
+        }
+
+        return $this;
+    }
+
+    public function addQueryParamString(string $value): static
+    {
+        if ($this->getQueryParamString() === '') {
+            $this->setQueryParamString($value);
+            return $this;
+        }
+
+        $this->queryParamString = $this->mergeQueryParamStrings(
+            $this->queryParamString,
+            $value
+        );
+
+        $this->parts = null;
+
+        if ($this->hasCleaned()) {
+            $this->cleanParts = $this->mergeQueryParamParts(
+                $this->getParts(),
+                $this->parseQueryParam($value),
+            );
+        } else {
+            $this->cleanParts = null;
+        }
+
+        return $this;
+    }
+
+    abstract protected function mergeQueryParamStrings(
+        string $queryParamString1,
+        string $queryParamString2
+    ): string;
+
+    abstract protected function mergeQueryParamParts(
+        array $queryParamParts1,
+        array $queryParamParts2,
+    ): array;
+
+    abstract protected function buildQueryParamStringFromParts(
+        array $queryParamParts,
+    ): array;
 
     public function getParts(): array
     {
@@ -55,6 +125,11 @@ abstract class AbstractQueryParam
      *  calling clean will be discarted.
      */
     abstract public function clean(callable $validate, bool $reset = false): void;
+
+    public function hasCleaned(): bool
+    {
+        return ($this->cleanParts !== null);
+    }
 
     public function __toString(): string
     {